pyproject.toml
[project]
name = "fastapi-sqlalchemy2-alembic-postgresql"
version = "0.1.0"
description = ""
authors = [
    {name = "Vitaliy Stepin"}
]
readme = "README.md"
requires-python = ">=3.13 <=3.14"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.1.0)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.0,<0.29.0)",
    "faker (>=40.1.2,<41.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
